+++ react/BlockChallenge.tsx
@@ -1,15 +1,68 @@
 import React from 'react'
-import getProfile from './graphql/getProfile.graphql'
-import { ExtensionPoint } from 'vtex.render-runtime'
-import { useQuery } from 'react-apollo'
-import { isNil, path } from 'ramda'
+import {
+  ExtensionPoint,
+  Session,
+  SessionResponse,
+  SessionUnauthorized,
+  SessionForbidden,
+} from 'vtex.render-runtime'
+
+import useSessionResponse from './hooks/useSessionResponse'
+
+const isSessionUnauthorized = (
+  session: SessionResponse | undefined
+): session is SessionUnauthorized =>
+  (session as SessionUnauthorized)?.type?.toLowerCase() === 'unauthorized'
+
+const isSessionForbidden = (
+  session: SessionResponse | undefined
+): session is SessionForbidden =>
+  (session as SessionForbidden)?.type?.toLowerCase() === 'forbidden'
+
+const isProfileAllowed = (sessionResponse: SessionResponse | undefined) => {
+  if (sessionResponse == null) {
+    return null
+  }
+
+  const hasAccessToTradePolicy = (sessionResponse as Session).namespaces?.store
+    ?.channel
+
+  const isLoggedIn = (sessionResponse as Session).namespaces?.profile?.email
+
+  if (hasAccessToTradePolicy) {
+    return 'authorized'
+  }
+
+  if (isLoggedIn) {
+    return 'forbidden'
+  }
+
+  return 'unauthorized'
+}
 
 const BlockChallenge = () => {
-  const { data: profileData, error, loading } = useQuery(getProfile, { ssr: false })
+  const sessionResponse = useSessionResponse()
 
-  if (loading || error || isNil(path(['profile'], profileData))) {
+  const isUnauthorized = isSessionUnauthorized(sessionResponse)
+  const isForbidden = isSessionForbidden(sessionResponse)
+  const profileCondition = isProfileAllowed(sessionResponse)
+
+  if (!sessionResponse) {
+    return null
+  }
+
+  const defaultHidden = sessionResponse == null
+
+  if (
+    defaultHidden ||
+    isUnauthorized === true ||
+    isForbidden === true ||
+    profileCondition === 'unauthorized' ||
+    profileCondition === 'forbidden'
+  ) {
     return <ExtensionPoint id="challenge-fallback" />
   }
+
   return <ExtensionPoint id="challenge-content" />
 }