+++ react/BlockContent.tsx
@@ -5,6 +5,7 @@ const CSS_HANDLES = ['challengeContentWrapper'] as const
 
 const BlockContent: FC<ChallengeProps> = ({ children }: ChallengeProps) => {
   const handles = useCssHandles(CSS_HANDLES)
+
   return (
     <div className={`mw9 center ${handles.challengeContentWrapper}`}>
       {children}