+++ CHANGELOG.md
@@ -7,15 +7,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added trade policy condition check
+
 ## [1.0.2] - 2020-06-23
 
 ## [1.0.1] - 2020-05-01
+
 ### Fixed
+
 - Functioning for SSRed stores.
 
 ## [1.0.0] - 2020-04-03
+
 ### Added
+
 - Added initial implementation
 
 ### Fixed
+
 - Used `profile` query check whether user is logged in